backend/requirements.txt
# Core
wheel==0.42.0
setuptools==69.0.0
numpy==1.24.3

# LangChain ecosystem (compatible versions)
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
langchain-core==0.1.23
langchain-text-splitters==0.0.1

# OpenAI
openai==1.10.0

# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
pydantic==2.5.3

# Vector store and document processing
faiss-cpu==1.7.4
pypdf==3.17.4
python-docx==1.1.0

# Speech-to-Text (Whisper)
llvmlite==0.41.1
numba==0.58.1
torch==2.1.2
regex==2023.12.25
tiktoken==0.5.2
openai-whisper==20231117

# Text-to-Speech
pyttsx3==2.90

# Utilities
python-dotenv==1.0.1